requirements.txt
openai
python-dotenv
playwright
langchain-openai
AzureOpenAI